# Improve time execution of AppByTokenLoader 