I want to merge this change to improve the execution time of AppByTokenLoader by cache usage. 

### Performance test 1
Only one token and only one app. 
Dataloader execution time in seconds. 
 |    | Old      | New      |
|----|----------|----------|
| 1  | 0.132958 | 0.126781 |
| 2  | 0.095072 | 0.021891 |
| 3  | 0.093437 | 0.017604 |
| 4  | 0.093917 | 0.015193 |
| 5  | 0.09317  | 0.014937 |
| 6  | 0.092564 | 0.016944 |
| 7  | 0.0908   | 0.017283 |
| 8  | 0.092013 | 0.014948 |
| 9  | 0.095801 | 0.017418 |
| 10 | 0.093705 | 0.019095 |

Result summary
|                            | Old        | New        | Diff       | Diff(%) |
|----------------------------|------------|------------|------------|---------|
| average                    |  0,0973437 |  0,0282094 |  0,0691343 |  71,02% |
| median                     | 0,09357100 | 0,01735050 | 0,07622050 |  81,46% |
| average without first call | 0,09338656 | 0,01725700 | 0,07612956 |  81,52% |
| median without first call  | 0,09343700 | 0,01728300 | 0,07615400 |  81,50% |

### Performance test 2
Only one app, but this app has 20 tokens with the same last 4 characters of the token.
To create such tokens I use this code:
```python
from saleor.app.models import App
token = "<EXISTING_TOKEN>"
app = App.objects.get(id="<APP_ID>")

for i in range(20):
    new_token = f"{i:02d}{token[2:]}"
    app.tokens.create(
        name=f"test_token{i}",
        auth_token=new_token,
    )
```
Dataloader execution time in seconds
|    | Old      | New      |
|----|----------|----------|
| 1  | 1.604753 | 1.622289 |
| 2  | 1.577014 | 0.017716 |
| 3  | 1.562331 | 0.018136 |
| 4  | 1.584505 | 0.012235 |
| 5  | 1.558621 | 0.017058 |
| 6  | 1.559019 | 0.015981 |
| 7  | 1.555267 | 0.016527 |
| 8  | 1.556953 | 0.015617 |
| 9  | 1.56113  | 0.011402 |
| 10 | 1.564756 | 0.017151 |

Result summary
|                            | Old        | New        | Diff       | Diff(%) |
|----------------------------|------------|------------|------------|---------|
| average                    |  1,5684349 |  0,1764112 |  1,3920237 |  88,75% |
| median                     | 1,56173050 | 0,01679250 | 1,54493800 |  98,92% |
| average without first call | 1,56439956 | 0,01575811 | 1,54864144 |  98,99% |
| median without first call  | 1,56113000 | 0,01652700 | 1,54460300 |  98,94% |



<!-- Please mention all relevant issue numbers. -->
<!-- GitHub issue number is required for external contributions. -->

# Impact

- [ ] New migrations
- [ ] New/Updated API fields or mutations
- [ ] Deprecated API fields or mutations
- [ ] Removed API types, fields, or mutations

# Docs

<!-- Docs are stored in a separate repository: https://github.com/saleor/saleor-docs/. -->
<!-- Please provide a link to the PR that updates documentation for your changes. -->
<!-- If changes in docs are not required, please mention that in the description. -->

- [ ] Link to documentation:

# Pull Request Checklist

<!-- Please keep this section. It will make the maintainer's life easier. -->

- [ ] Privileged queries and mutations are either absent or guarded by proper permission checks
- [x] Database queries are optimized and the number of queries is constant
- [ ] Database migrations are either absent or optimized for zero downtime
- [x] The changes are covered by test cases
- [ ] All new fields/inputs/mutations have proper labels added (`ADDED_IN_X`, `PREVIEW_FEATURE`, etc.)
- [ ] All migrations have proper dependencies
- [ ] All indexes are added concurrently in migrations
- [ ] All RunSql and RunPython migrations have revert option defined